requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
databases[postgresql]
psycopg2-binary
openai
requests
pinecone
python-dotenv